fix double

+++ peripheral/stm32g4/hrpwm.cpp
@@ -63,7 +63,7 @@ void HRPWM::set_frequency_hz(uint32_t frequency_hz, uint16_t min_off_ns, uint16_
     }
     
     count_per_ns_ = CPU_FREQUENCY_HZ * prescaler_/ 4 / 1.e9; // Datasheet says /8 not /4, but /4 seems to give correct scale
-    period_ = (double) CPU_FREQUENCY_HZ*prescaler_/2/frequency_hz;
+    period_ = (float) CPU_FREQUENCY_HZ*prescaler_/2/frequency_hz;
     regs_.sTimerxRegs[ch_a_].PERxR = period_;
     regs_.sTimerxRegs[ch_b_].PERxR = period_;
     regs_.sTimerxRegs[ch_c_].PERxR = period_;
